Move all fields of the create course form from the course_creator component to the course_form component
